Reflect event store's change for position from int to long

+++ src/main/java/org/fuin/cqrs4j/ProjectionService.java
@@ -54,6 +54,6 @@ public interface ProjectionService {
      * @param nextEventNumber
      *            Number of the next event to read.
      */
-    public void updateProjectionPosition(@NotNull StreamId streamId, @NotNull Integer nextEventNumber);
+    public void updateProjectionPosition(@NotNull StreamId streamId, @NotNull Long nextEventNumber);
 
 }
